requirements.txt
edge_tts
pydub
SpeechRecognition
keyboard
PyAudio
setuptools
PyYAML
langchain
langchain_openai
langchain_community
wikipedia
python-dotenv
youtube_transcript_api
google-api-python-client